refactor(rel): extract value comparison semantics (#1164)
